+++ print_other.go
@@ -1,0 +1,7 @@
+//go:build !darwin
+
+package main
+
+// applyPrintOrientation is a no-op off macOS; other platforms' print
+// dialogs manage orientation themselves.
+func applyPrintOrientation(landscape bool) {}